put the cateogries and posts into a single blog file
