--- models/category.py
@@ -1,38 +1,0 @@
-import sqlite3
-from db import db
-
-
-class CategoryModel(db.Model):
-    __tablename__ = 'categories'
-
-    id = db.Column(db.Integer, primary_key=True)
-
-    post_id = db.Column(db.Integer, db.ForeignKey('posts.id'))
-    post = db.relationship('PostModel')
-    name = db.Column(db.String(80))
-
-    def __init__(self, name):
-        self.name = name
-
-    def json(self):
-        return {
-            'id': self.id,
-            'name': self.name,
-            'post_id': self.post_id
-        }
-
-    @classmethod
-    def find_by_name(cls, name):
-        return cls.query.filter_by(name=name).first()
-
-    @classmethod
-    def find_all(cls):
-        return cls.query.all()
-
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
