+++ resources/blog.py
@@ -1,7 +1,26 @@
 import sqlite3
 from flask_restful import Resource
 
-from models.category import CategoryModel
+from models.blog import PostModel, CategoryModel
+
+
+class Post(Resource):
+    def get(self, name):
+        post = PostModel.find_by_name(name)
+        if post:
+            return post.json()
+        else:
+            return {'message': 'Post not found'}, 404
+
+    # delete + JWT
+
+class PostList(Resource):
+    """Lists the items"""
+    def get(self):
+        posts = [post.json() for post in PostModel.find_all()]
+        return {
+            'posts': posts
+        }
 
 
 class Category(Resource):
